requirements.txt
Flask==1.1.2
flask-apispec==0.11.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
python-decouple==3.4
SQLAlchemy==1.3.23
gunicorn==20.0.4